requirement.txt
alembic==1.13.2
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
email-validator==2.2.0
flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib-metadata==8.0.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing-extensions==4.12.2
werkzeug==3.0.3
wtforms==3.1.2
zipp==3.19.2
